Move serialization interfaces from interfaces.ts to dataFormats.ts
Extracts IConversationData, ITermToSemanticRefIndexData, and
ITermToSemanticRefIndexItem into a new dataFormats.ts file. Re-exports
them from interfaces.ts for backward compatibility with existing callers.
